Implement query choice field with easyform
