+++ buildout.cfg
@@ -11,3 +11,4 @@ collective.elasticsearch = git git@github.com:collective/collective.elasticsearc
 collective.documentviewer = git git@github.com:collective/collective.documentviewer.git branch=master
 castle.theme = git git@github.com:castlecms/castle.theme.git branch=master
 collective.easyform = git git@github.com:collective/collective.easyform.git
+plone.supermodel = git git@github.com:plone/plone.supermodel.git
