+++ castle/cms/constants.py
@@ -7,3 +7,4 @@ ALL_USERS = '--USERS--'
 DEFAULT_SITE_LAYOUT_REGISTRY_KEY = 'castle.cms.mosaic.default_site_layout'
 CRAWLED_SITE_ES_DOC_TYPE = 'website'
 CRAWLED_DATA_KEY = 'castle.cms.crawldata'
+ANONYMOUS_USER = "Anonymous User"
